Revert "revamp 'config' module interface for accessing cross-module variables, integrate new serializers module from PygameCommunityBot (#4)"
This reverts commit 366012a382ce5883617ed57bd908ebec551a53f1.
