+++ requirements.txt
@@ -1,4 +1,2 @@
-black>=22.1
 discord.py>=1.7.3
 python-dotenv>=0.19
-typeguard>=2.13
