+++ snakecore/utils/__init__.py
@@ -9,21 +9,21 @@ This module defines some utility functionality for the library and general bot d
 from typing import Optional
 import discord
 from snakecore import config
-from snakecore.constants import UNSET, UNSET_TYPE
 from . import embed_utils
 from .utils import *
 
 
-def init(client: Union[UNSET_TYPE, discord.Client] = UNSET):
-    if not isinstance(client, (discord.Client, UNSET_TYPE)):
+def init(client: Optional[discord.Client] = None):
+    if not isinstance(client, (discord.Client, type(None))):
         raise TypeError(
             f"argument 'client' must be None or of type discord.Client,"
             f" not {client.__class__.__name__}"
         )
 
-    config.set_value("global_client", client, ignore_if_set=True)
-    config.set_value("utils_is_init", True)
+    if config.client is None:
+        config.client = client
+    config.utils_is_init = True
 
 
 def is_init() -> bool:
-    return config.get_value("utils_is_init", wanted_value_cls=bool)
+    return config.utils_is_init
